test: review datasync tests
